pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-search"
version = "2.0.0"
description = "A pluggable RAG-powered search engine with semantic caching and intelligent query routing"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["rag", "search", "llm", "semantic-cache", "vector-database"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "openai>=1.0.0",
    "qdrant-client>=1.6.0",
    "sentence-transformers>=2.2.0",
    "faiss-cpu>=1.7.0",
    "duckduckgo-search>=4.0.0",
    "python-dotenv>=1.0.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
web = [
    "streamlit>=1.28.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
all = [
    "agentic-search[web,dev]",
]

[project.scripts]
agentic-search = "agentic_search.ui.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/agentic-search"
Documentation = "https://github.com/yourusername/agentic-search#readme"
Repository = "https://github.com/yourusername/agentic-search"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py39"
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=agentic_search --cov-report=term-missing"
